# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "数智化管理系统",
    "description": "数智化管理系统 API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "账号密码登录",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "登出",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sso/qrcode/generate": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "生成 SSO 扫码二维码",
        "operationId": "sso_qrcode_generate_api_v1_auth_sso_qrcode_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SSOQRCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSOQRCodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sso/qrcode/check": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "轮询扫码状态",
        "description": "演示版：直接返回 waiting\n真实实现：查 Redis 中 qr_token 对应的扫码状态",
        "operationId": "sso_qrcode_check_api_v1_auth_sso_qrcode_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SSOCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSOCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/reset/request": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "发送重置密码验证码",
        "operationId": "password_reset_request_api_v1_auth_password_reset_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/reset/confirm": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "确认重置密码",
        "operationId": "password_reset_confirm_api_v1_auth_password_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "认证"
        ],
        "summary": "获取当前登录用户信息",
        "operationId": "get_me_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "account": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Account",
            "description": "手机号/邮箱/工号"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          },
          "rememberMe": {
            "type": "boolean",
            "title": "Rememberme",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "account",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin"
          },
          "userInfo": {
            "$ref": "#/components/schemas/UserInfo"
          }
        },
        "type": "object",
        "required": [
          "token",
          "refreshToken",
          "expiresIn",
          "userInfo"
        ],
        "title": "LoginResponse"
      },
...
      "PasswordResetConfirm": {
        "properties": {
          "account": {
            "type": "string",
            "title": "Account"
          },
          "verifyCode": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Verifycode"
          },
          "newPassword": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "account",
          "verifyCode",
          "newPassword"
        ],
        "title": "PasswordResetConfirm"
      },
      "PasswordResetRequest": {
        "properties": {
          "account": {
            "type": "string",
            "title": "Account"
          },
          "verifyType": {
            "type": "string",
            "pattern": "^(email|sms)$",
            "title": "Verifytype"
          }
        },
        "type": "object",
        "required": [
          "account",
          "verifyType"
        ],
        "title": "PasswordResetRequest"
      },
...
      "SSOCheckRequest": {
        "properties": {
          "qrToken": {
            "type": "string",
            "title": "Qrtoken"
          }
        },
        "type": "object",
        "required": [
          "qrToken"
        ],
        "title": "SSOCheckRequest"
      },
      "SSOCheckResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "userInfo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SSOCheckResponse"
      },
      "SSOQRCodeRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "pattern": "^(wechat|dingtalk|feishu)$",
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "SSOQRCodeRequest"
      },
      "SSOQRCodeResponse": {
        "properties": {
          "qrToken": {
            "type": "string",
            "title": "Qrtoken"
          },
          "qrImageUrl": {
            "type": "string",
            "title": "Qrimageurl"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin"
          },
          "pollUrl": {
            "type": "string",
            "title": "Pollurl"
          }
        },
        "type": "object",
        "required": [
          "qrToken",
          "qrImageUrl",
          "expiresIn",
          "pollUrl"
        ],
        "title": "SSOQRCodeResponse"
      },
...
      "UserInfo": {
        "properties": {
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "userId",
          "name",
          "role"
        ],
        "title": "UserInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}